Refined app permissions
Added SLACK_APP_TOKEN and SLACK_BOT_TOKEN in the docker-compose file as passthrough. At the same time, disable the "messages_tab" in the app_home feature of the manifest file. Removal of certain bot events and scopes were also made to limit the Slack bot's functionality mainly to writing to public chats.