Remove 'exclude_dir' argument from Docker configurations
The Docker configuration files, docker-compose.yml and smib-fast.Dockerfile, have been modified to remove the 'exclude_dir' argument. This argument was previously used to exclude directories from the Docker build process, however, this exclusion was found unnecessary. Now, the Docker build workflow is simpler and does not exclude any directories.

This is now done in the load_injection_container calls to properly sort the injection issues