Handle SlackApiError exceptions in the space open/close plugin
Added error handling for SlackApiError exceptions within the `space_open` and `space_closed` methods in the space open/close plugin. This commit aims to prevent any interruptions in communication with the slack API. Also, the default value of `SPACE_OPEN_ANNOUNCE_CHANNEL_ID` was changed to 'space-open-announce' in the config file.