Update README with file location details and Docker config changes
The commit adds explicit file location details for 'logging.json' and '.env' in the README. The Docker volume paths have been updated in the Linux instructions. In addition, it provides an important note on mapping '/app/config' to a host directory which mandates adding the 2 external files to this location.