feat: Update Docker Compose files to allow building with different branches
Changed Docker Compose files to specify the branch of the GitHub repository
being used for the build.
Add a new `docker-compose-branch.yml` file that allows a default branch of `master`
to be used, or a specific branch to be specified.
Also updated the README file with instructions on how to specify the branch
for the build.